backend/services/question-service/requirements.txt
config~=0.5.1
python-dotenv~=1.1.1